Move WorkSubTypes component into Edit elements for reuse with collections
