+++ app/components/edit/work_subtypes_component.html.erb
@@ -3,7 +3,7 @@
     <%= content %>
   <% else %>
     <%= render Elements::Forms::FieldsetComponent.new(label:, tooltip:, aria: fieldset_aria) do %>
-      <%= tag.p class: 'form-text mb-3' do %><%= help_text %><% end %>
+      <p class="mb-3"><%= help_text %></p>
       <%= render Elements::Forms::InvalidFeedbackComponent.new(form:, field_name: error_field_name) %>
 
       <% if required_subtypes.present? %>
